Add files via upload
This is the FAQ ''page''. It was added onto the index.blade.php file on our PCs but to avoid any conflict with the GitHub, I've uploaded it under the name faq.blade.php instead. just be aware this is the exact same as the home page until we sort out in future iterations what pages need to be seperate from the home page etc.